pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-lab"
version = "0.1.0"
description = "Machine Learning experiments with AI agents and MCP"
requires-python = ">=3.11"
dependencies = [
    "accelerate",
    "asyncio",
    "ipykernel",
    "huggingface_hub[cli]",
    "jupyter",
    "mcp",
    "mlx",
    "mlx-lm",
    "mlx-vlm",
    "nest-asyncio",
    "openai",
    "Pillow",
    "strands-agents[a2a]",
    "strands-agents-tools",
    "torch",
    "torchvision",
    "transformers",
]

[project.scripts]
mcp-server = "mcp_lab.lab_server.main:main"
a2a-server = "a2a_lab.server:main"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["src/a2a_lab", "src/mcp_lab"]

[tool.uv.sources]